requirements.txt
# whisper-diarization
nemo_toolkit[asr]>=2.3.0
nltk
faster-whisper>=1.1.0
git+https://github.com/MahmoudAshraf97/demucs.git
git+https://github.com/oliverguhr/deepmultilingualpunctuation.git
git+https://github.com/MahmoudAshraf97/ctc-forced-aligner.git
onnxruntime
torchaudio
transformers
# emo classification with Speech2Emo
librosa
# emo classification with Emotion2Vec
funasr
